+++ src/sample.cpp
@@ -6,7 +6,7 @@ void Sample::ApplyProperties() {
     throw std::runtime_error("Base sample class init shouldn't be called");
 }
 
-float Sample::GetSample(float time) {
+float Sample::GetSample(double time) {
     throw std::runtime_error("Base sample class GetSample shouldn't be called");
     return 0;
 }
